Tweaked style and refactored

+++ test/fib.cpp
@@ -1,8 +1,8 @@
 #include <cassert>
 #include "kgen.hpp"
 
-class fib_gen : public kgen::gen<int, 2> {
-    int next() override {
+class fib_gen : public kgen::gen<unsigned long long, 2> {
+    unsigned long long next() override {
         return prev(-2) ? prev(-1) + prev(-2) : 1;
     }
 };
